+++ lib/spior/iptables.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Spior
   module Iptables
   end
@@ -6,3 +8,4 @@ end
 require_relative 'iptables/root'
 require_relative 'iptables/tor'
 require_relative 'iptables/default'
+require_relative 'iptables/rules'