+++ lib/spior/iptables/root.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'interfacez'
 
 module Spior
@@ -6,8 +8,8 @@ module Spior
       def initialize
         @lo      = Interfacez.loopback
         @lo_addr = Interfacez.ipv4_address_of(@lo)
-        @i = Helpers::Exec.new("iptables")
-        Spior::Copy.new.save
+        @i = Helpers::Exec.new('iptables')
+        @debug = false
       end
 
       def run!
@@ -22,63 +24,60 @@ module Spior
       end
 
       def stop!
-        ipt "-F"
-        ipt "-X"
-        ipt "-t nat -F"
-        ipt "-t nat -X"
-        ipt "-t mangle -F"
-        ipt "-t mangle -X"
+        Msg.p 'Clearing Iptables rules...'
+        ipt '-F'
+        ipt '-X'
+        ipt '-t nat -F'
+        ipt '-t nat -X'
+        ipt '-t mangle -F'
+        ipt '-t mangle -X'
       end
 
       private
 
       def ipt(line)
         @i.run("#{line}")
-        puts "added - #{@i} #{line}"
+        puts "Added - iptables #{line}" if @debug
       end
 
-      def redirect
-      end
+      def redirect; end
 
-      def input
-      end
+      def input; end
 
-      def output
-      end
+      def output; end
 
-      def all
-      end
+      def all; end
 
       def bogus_tcp_flags
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags FIN,SYN,RST,PSH,ACK,URG NONE -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags FIN,SYN FIN,SYN -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags SYN,RST SYN,RST -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags FIN,RST FIN,RST -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags FIN,ACK FIN -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ACK,URG URG -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ACK,FIN FIN -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ACK,PSH PSH -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ALL ALL -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ALL NONE -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ALL FIN,PSH,URG -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ALL SYN,FIN,PSH,URG -j DROP"
-        ipt "-t mangle -A PREROUTING -p tcp --tcp-flags ALL SYN,RST,ACK,FIN,URG -j DROP"
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags FIN,SYN,RST,PSH,ACK,URG NONE -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags FIN,SYN FIN,SYN -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags SYN,RST SYN,RST -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags FIN,RST FIN,RST -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags FIN,ACK FIN -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ACK,URG URG -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ACK,FIN FIN -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ACK,PSH PSH -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ALL ALL -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ALL NONE -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ALL FIN,PSH,URG -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ALL SYN,FIN,PSH,URG -j DROP'
+        ipt '-t mangle -A PREROUTING -p tcp --tcp-flags ALL SYN,RST,ACK,FIN,URG -j DROP'
       end
 
       def bad_packets
         # new packet not syn
-        ipt "-t mangle -A PREROUTING -p tcp ! --syn -m conntrack --ctstate NEW -j DROP"
+        ipt '-t mangle -A PREROUTING -p tcp ! --syn -m conntrack --ctstate NEW -j DROP'
         # fragment  packet
-        ipt "-A INPUT -f -j DROP"
+        ipt '-A INPUT -f -j DROP'
         # XMAS
-        ipt "-A INPUT -p tcp --tcp-flags ALL ALL -j DROP"
+        ipt '-A INPUT -p tcp --tcp-flags ALL ALL -j DROP'
         # null packet
-        ipt "-A INPUT -p tcp --tcp-flags ALL NONE -j DROP"
+        ipt '-A INPUT -p tcp --tcp-flags ALL NONE -j DROP'
       end
 
       def spoofing
-        subs=["224.0.0.0/3", "169.254.0.0/16", "172.16.0.0/12", "192.0.2.0/24", "0.0.0.0/8", "240.0.0.0/5"]
-        subs.each do |sub|
+        subs = %w[224.0.0.0/3 169.254.0.0/16 172.16.0.0/12 192.0.2.0/24 0.0.0.0/8 240.0.0.0/5]
+        subs.map do |sub|
           ipt "-t mangle -A PREROUTING -s #{sub} -j DROP"
         end
         ipt "-t mangle -A PREROUTING -s #{@lo_addr}/8 ! -i #{@lo} -j DROP"