+++ lib/spior/service/start.rb
@@ -1,26 +1,14 @@
-require 'tty-which'
+# frozen_string_literal: true
 
 module Spior
   module Service
     module_function
 
+    # Service.start should start Tor if not alrealy running
+    # And start to redirect the local traffic with Iptables
     def start
-      if TTY::Which.exist?('systemctl')
-        state = `systemctl is-active tor`.chomp
-        unless state == 'active'
-          Helpers::Exec.new("systemctl").run("start tor")
-          Msg.p "TOR started."
-        end
-      elsif TTY::Which.exist? 'sv'
-        unless File.exist? '/var/service/tor'
-          Helpers::Exec.new('ln').run('-s /etc/sv/tor /var/service/tor')
-          Msg.p "TOR started."
-        end
-      elsif File.exist? '/etc/init.d/tor'
-        Helpers::Exec.new('/etc/init.d/tor').run('start')
-      else
-        Msg.report "Don't known yet how to start Tor for your system."
-      end
+      Tor.start
+      Iptables::Tor.new.run!
     end
   end
 end