+++ lib/spior/tor.rb
@@ -1,6 +1,13 @@
+# frozen_string_literal: true
+
 module Spior
+
+  # The module Tor interract with Tor on your system.
   module Tor
   end
 end
 
-require_relative 'tor/info'
+require_relative 'tor/data'
+require_relative 'tor/config'
+require_relative 'tor/start'
+require_relative 'tor/stop'